+++ R/setup.R
@@ -23,7 +23,9 @@ update_rextendr_version <- function(desc_path, cur_version = NULL) {
   cur <- cur_version %||% as.character(utils::packageVersion("rextendr"))
   prev <- rextendr_version(desc_path = desc_path)
 
-  if (!is.na(cur) && !is.na(prev) && package_version(cur) < package_version(prev)) {
+  if (
+    !is.na(cur) && !is.na(prev) && package_version(cur) < package_version(prev)
+  ) {
     cli::cli_alert_warning(c(
       "Installed rextendr is older than the version used with this package",
       "You have {.str {cur}} but you need {.str {prev}}"
@@ -34,8 +36,10 @@ update_rextendr_version <- function(desc_path, cur_version = NULL) {
 }
 
 update_sys_reqs <- function(desc_path) {
-  cur <- "Cargo (Rust's package manager), rustc"
-  prev <- stringi::stri_trim_both(desc::desc_get("SystemRequirements", file = desc_path)[[1]])
+  cur <- "Cargo (Rust's package manager), rustc >= 1.65.0, xz"
+  prev <- stringi::stri_trim_both(
+    desc::desc_get("SystemRequirements", file = desc_path)[[1]]
+  )
 
   if (is.na(prev)) {
     update_description("SystemRequirements", cur, desc_path = desc_path)
@@ -50,10 +54,14 @@ update_sys_reqs <- function(desc_path) {
 }
 
 update_description <- function(field, value, desc_path) {
-  cli::cli_alert_info("Setting {.var {field}} to {.str {value}} in the {.file DESCRIPTION} file.")
+  cli::cli_alert_info(
+    "Setting {.var {field}} to {.str {value}} in the {.file DESCRIPTION} file."
+  )
   desc::desc_set(field, value, file = desc_path)
 }
 
 rextendr_version <- function(desc_path = ".") {
-  stringi::stri_trim_both(desc::desc_get("Config/rextendr/version", desc_path)[[1]])
+  stringi::stri_trim_both(desc::desc_get("Config/rextendr/version", desc_path)[[
+    1
+  ]])
 }